05e2091ca24fd3492aadd7e3c1aacf75dfb6b873 merge bitcoin#30464: Fix MSVC warning C4101 "unreferenced local variable" (Kittywhiskers Van Gogh)
f00dc7813419d0dd6d09d405dc44b8be2d2e3cd6 build: drop `-Wdangling-reference` GCC suppression (Kittywhiskers Van Gogh)
6d81d06090fb584fb5d216de3fc012ae78c87637 merge bitcoin#27605: Replace global find_value function with UniValue::find_value method (Kittywhiskers Van Gogh)
203a1c17a84e20dcddf3fde2969014cd276754e7 merge bitcoin#26612: pass named argument value as string_view (Kittywhiskers Van Gogh)
41eb3b05c7a144901ea3bd1311537d294a992128 merge bitcoin#25714: Avoid std::string copies (Kittywhiskers Van Gogh)
6f03a13e577b42b8a411b79baa31b6929c17a996 merge bitcoin#25736: Remove unused and confusing set*() return value (Kittywhiskers Van Gogh)
27cffb13e8da95be316bf519a4594646d1803f6b refactor: shed avoidable casts in Dash-specific UniValue-ret functions (Kittywhiskers Van Gogh)
32958daf50306db2de4fcbe9624398a3c20aaadf merge bitcoin#25611: Avoid brittle, narrowing and verbose integral type confusions (Kittywhiskers Van Gogh)
7802752bd5a0fe028aedc43455318a95b7d3e912 merge bitcoin#25617: univalue test cleanups (Kittywhiskers Van Gogh)
94bbbf1b4fc323d833ad1aeb511c2f71996ec675 partial bitcoin#25551: Throw exception on invalid Univalue pushes over silent ignore (Kittywhiskers Van Gogh)
d980870301f3d060cf040825e8f76a53a090f039 revert: revert bitcoin#25464 (Reduce Univalue push_backV peak memory usage in listtransactions) (Kittywhiskers Van Gogh)
1e023e5995853737be1935247a6d421f3d7da413 merge bitcoin#25369: Unsubtree Univalue (Kittywhiskers Van Gogh)
6b9e682f8cbe5b18e296dc9485c3144434cf9660 merge bitcoin#25249: Bump univalue subtree (Kittywhiskers Van Gogh)
b3443eebd49919287453b303270ba2a044537e9e Squashed 'src/univalue/' changes from 6c19d050a9..de4f73ddca (Kittywhiskers Van Gogh)
9db7b12119256e74f422892dee90c89ae87daba9 merge bitcoin#25113: Bump univalue subtree (Kittywhiskers Van Gogh)
88f794f4a1e330ad4f10d0c03631ed473646951a Squashed 'src/univalue/' changes from a44caf65fe..6c19d050a9 (Kittywhiskers Van Gogh)
b07e5506e8bc2dec275e283f50c2661f5ec4d216 merge bitcoin#25153: Use getInt<T> over get_int/get_int64 (Kittywhiskers Van Gogh)
b1ae6f5f59b7a45d72b4fa14087a07818da8a0f3 Squashed 'src/univalue/' changes from 2740c4f712..a44caf65fe (Kittywhiskers Van Gogh)
1780961a6152ccb109f9c51bf54dd7209df9b8f6 revert: bitcoin#25464 (Reduce Univalue push_backV peak memory usage in listtransactions) (Kittywhiskers Van Gogh)

Pull request description:

  ## Motivation

  To get rid of the `-Wdangling-reference` suppression ([source](https://github.com/dashpay/dash/blob/f648039258d2192ec5e505721e34edaa0bda568c/configure.ac#L461-L467)) for GCC introduced in [dash#6639](https://github.com/dashpay/dash/pull/6639) by backporting the set of changes needed to backport [bitcoin#27605](https://github.com/bitcoin/bitcoin/pull/27605).

  ## Additional Information

  * At some point, the contents on `univalue/` diverged from official repository ([bitcoin-core/univalue-subtree](https://github.com/bitcoin-core/univalue-subtree)), which makes updating the subtree a more difficult prospect.

    We realign the subtree by updating it to [`a44caf65`](https://github.com/bitcoin-core/univalue-subtree/commit/a44caf65) (last sane commit, taken from [bitcoin#22646](https://github.com/bitcoin/bitcoin/pull/22646), which was backported as [dash#4823](https://github.com/dashpay/dash/pull/4823)) and marking all merge conflicts in favour of the incoming update. This lets us do the final few subtree updates before it's no longer treated as a subtree.

  * To allow for subtree updates, [bitcoin#25464](https://github.com/bitcoin/bitcoin/pull/25464) (backported as 001d7bad0e in [dash#6698](https://github.com/dashpay/dash/pull/6698)) was temporarily reverted and then reapplied _after_ the directory was unsubtree'd.

  * [bitcoin#30464](https://github.com/bitcoin/bitcoin/pull/30464) needed to be backported to keep the whitespace linter happy (failed [build](https://github.com/dashpay/dash/actions/runs/16463979981/job/46536933377#step:5:119)).

  ## Breaking Changes

  None expected.

  ## Checklist

  - [x] I have performed a self-review of my own code
  - [x] I have commented my code, particularly in hard-to-understand areas **(note: N/A)**
  - [x] I have added or updated relevant unit/integration/functional/e2e tests
  - [x] I have made corresponding changes to the documentation **(note: N/A)**
  - [x] I have assigned this pull request to a milestone _(for repository code-owners and collaborators only)_

ACKs for top commit:
  UdjinM6:
    utACK 05e2091ca24fd3492aadd7e3c1aacf75dfb6b873

Tree-SHA512: fdd35699defcc24a396fc9746a8eca26a5866666e9ae049f7d9e38fe84021aa49169e48b9408d3c47c6cd41aa9523f0f94a8b91ecc874a962bf5e6005d7eff21

+++ src/rpc/masternode.cpp
@@ -175,7 +175,7 @@ static RPCHelpMan masternode_status()
         mnObj.pushKV("proTxHash", dmn->proTxHash.ToString());
         mnObj.pushKV("type", std::string(GetMnType(dmn->nType).description));
         mnObj.pushKV("collateralHash", dmn->collateralOutpoint.hash.ToString());
-        mnObj.pushKV("collateralIndex", (int)dmn->collateralOutpoint.n);
+        mnObj.pushKV("collateralIndex", dmn->collateralOutpoint.n);
         mnObj.pushKV("dmnState", dmn->pdmnState->ToJson(dmn->nType));
     }
     mnObj.pushKV("state", node.mn_activeman->GetStateString());
